# jaxrs-spec-generated/swagger.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "SOTE API",
    "description" : "SOTE API",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "/v1"
  } ],
  "paths" : {
    "/contents" : {
      "get" : {
        "tags" : [ "Contents" ],
        "summary" : "Lists contents",
        "description" : "Lists contents\n",
        "operationId" : "listContents",
        "parameters" : [ {
          "name" : "parentId",
          "in" : "query",
          "description" : "Filter results by parent id",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "type",
          "in" : "query",
          "description" : "Filter results by content type",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }, {
          "name" : "categorySlug",
          "in" : "query",
          "description" : "Filter by category slug",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a list of contents",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Content"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/contents/{contentId}" : {
      "get" : {
        "tags" : [ "Contents" ],
        "summary" : "Finds content",
        "description" : "Finds single content\n",
        "operationId" : "findContent",
        "parameters" : [ {
          "name" : "contentId",
          "in" : "path",
          "description" : "content id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a single content",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Content"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/contents/{contentId}/images" : {
      "get" : {
        "tags" : [ "Contents" ],
        "summary" : "Returns a list of content images",
        "description" : "Returns a list of content images\n",
        "operationId" : "listContentImages",
        "parameters" : [ {
          "name" : "contentId",
          "in" : "path",
          "description" : "Content id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "type",
          "in" : "query",
          "description" : "Filter by type",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns list of content images",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ContentImage"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/contents/{contentId}/images/{imageId}" : {
      "get" : {
        "tags" : [ "Contents" ],
        "summary" : "Returns a single content image",
        "description" : "Returns a single content image\n",
        "operationId" : "findContentImage",
        "parameters" : [ {
          "name" : "contentId",
          "in" : "path",
          "description" : "Content Id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "imageId",
          "in" : "path",
          "description" : "Content image id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns an content image",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ContentImage"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/contents/{contentId}/images/{imageId}/data" : {
      "get" : {
        "tags" : [ "Contents" ],
        "summary" : "Returns an content image data",
        "description" : "Returns an content image data\n",
        "operationId" : "getContentImageData",
        "parameters" : [ {
          "name" : "contentId",
          "in" : "path",
          "description" : "Content id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "imageId",
          "in" : "path",
          "description" : "Content image id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "size",
          "in" : "query",
          "description" : "Maximum width or height of image",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns an content image data",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "type" : "string",
                  "format" : "binary"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/octet-stream" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
    "/categories" : {
      "get" : {
        "tags" : [ "Categories" ],
        "summary" : "Lists categories",
        "description" : "Lists categories\n",
        "operationId" : "listCategories",
        "parameters" : [ {
          "name" : "firstResult",
          "in" : "query",
          "description" : "First result",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "maxResults",
          "in" : "query",
          "description" : "Max results",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Returns a list of categories",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Category"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request was sent to the server",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "403" : {
            "description" : "Attempted to make a call with unauthorized client",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Forbidden"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "application/json;charset=utf-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InternalServerError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Content" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "parentId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "level" : {
            "type" : "integer",
            "format" : "int32"
          },
          "slug" : {
            "type" : "string"
          },
          "category" : {
            "type" : "string",
            "description" : "Category slug"
          },
          "type" : {
            "type" : "string",
            "description" : "Content type",
            "enum" : [ "PAGE", "NEWS", "LINK" ]
          },
          "title" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LocalizedValue"
            }
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LocalizedValue"
            }
          },
          "created" : {
            "type" : "string",
            "description" : "Create time.",
            "format" : "date-time"
          },
          "modified" : {
            "type" : "string",
            "description" : "Create time.",
            "format" : "date-time"
          }
        }
      },
      "ContentImage" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "string"
          },
          "size" : {
            "type" : "integer",
            "format" : "int64"
          },
          "type" : {
            "type" : "string"
          }
        }
      },
...
      "Category" : {
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "slug" : {
            "type" : "string"
          },
          "title" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/LocalizedValue"
            }
          }
        }
      },
...
      "LocalizedValue" : {
        "type" : "object",
        "properties" : {
          "language" : {
            "type" : "string"
          },
          "value" : {
            "type" : "string"
          }
        }
      },
...
      "InternalServerError" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "Forbidden" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
      "BadRequest" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}